update NNPACK convolution
